flask-app/requirements.txt
flask
flask-wtf

modzy-sdk>=0.11.0
opencv-python-headless
pandas
Jetson.GPIO
